% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParConfig.R
\name{downloadParConfig}
\alias{downloadParConfig}
\title{Downloads a single ParConfig.}
\usage{
downloadParConfig(id)
}
\arguments{
\item{id}{[\code{character}]
Unique identifier for a Parameter Set.
This will be downloaded from the mlrHyperopt servers.}
}
\value{
[\code{ParConfig}]
}
\description{
Retrieve a Paramater Configuration for the given id from the mlrHyperopt servers.
}
\examples{
par.config = downloadParConfig("8")
print(par.config)
}

