% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParConfig.R
\name{downloadParConfigs}
\alias{downloadParConfigs}
\title{Downloads multiple Parameter Configurations from the mlrHyperopt servers}
\usage{
downloadParConfigs(ids)
}
\arguments{
\item{ids}{[\code{character}]
One ore more unique identifiers of the Parameter Set}
}
\value{
[List of \code{ParConfig}s]
}
\description{
Retrieve the Paramater Configurations for the given ids from the mlrHyperopt servers.
}

