% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParSet.R
\name{downloadParSet}
\alias{downloadParSet}
\title{Downloads a Parameter Set from the mlrHyperopt servers}
\usage{
downloadParSet(ids)
}
\arguments{
\item{ids}{[\code{character}]
One ore more unique identifiers of the Parameter Set}
}
\value{
[\code{ParamSet}|list of ParamSets]
}

