% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadParSets.R
\name{downloadParSets}
\alias{downloadParSets}
\title{Downloads a Parameter Set from the mlrHyperopt servers}
\usage{
downloadParSets(ids)
}
\arguments{
\item{ids}{[\code{character}]
One ore more unique identifiers of the Parameter Set}
}
\value{
[List of \code{ParamSets}]
}

