% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHyperControl.R
\name{generateHyperControl}
\alias{generateHyperControl}
\title{Generates a hyperparameter tuning control object}
\usage{
generateHyperControl(task, learner, par.config)
}
\arguments{
\item{task}{[\code{Task}]
Task}

\item{learner}{[\code{Learner}]
Learner}

\item{par.config}{[\code{ParConfig}]}
}
\value{
[\code{HyperControl}]
}
\description{
Tries to automatically create a suitable hyperparameter tuning control.
}

