% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultParConfig.R
\name{getDefaultParConfig}
\alias{getDefaultParConfig}
\title{Gives a default ParConfig for a Learner}
\usage{
getDefaultParConfig(learner, drop.par.vals = TRUE)
}
\arguments{
\item{learner}{[\code{Learner}]
An mlr Learner.}

\item{drop.par.vals}{[\code{logical(1)}]
Should par.vals set in the learner be dropped if they are part of the tuning par.set?
Default is \code{TRUE}.}
}
\value{
[\code{ParConfig}]
}
\description{
For a given Learner this method will return a useful default Parameter Set for Tuning.
}
\examples{
learner = makeLearner("classif.randomForest")
par.config = getDefaultParConfig(learner)
print(par.config)
}
