% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{getHyperControlMeasures}
\alias{getHyperControlMeasures}
\title{Get the Measures}
\usage{
getHyperControlMeasures(hyper.control)
}
\arguments{
\item{hyper.control}{[\code{\link{HyperControl}}]
The Hyperparameter Control Object}
}
\value{
[\code{\link[mlr]{measures}}].
}
\description{
Get the Measures
}
\seealso{
Other HyperControl: 
\code{\link{getHyperControlMlrControl}()},
\code{\link{getHyperControlResampling}()},
\code{\link{makeHyperControl}()},
\code{\link{setHyperControlMeasures}()},
\code{\link{setHyperControlMlrControl}()},
\code{\link{setHyperControlResampling}()}
}
\concept{HyperControl}
