% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{getHyperControlMlrControl}
\alias{getHyperControlMlrControl}
\title{Get the mlr Tuning Object}
\usage{
getHyperControlMlrControl(hyper.control)
}
\arguments{
\item{hyper.control}{[\code{\link{HyperControl}}]
The Hyperparameter Control Object}
}
\value{
[\code{\link[mlr]{TuneControl}}].
}
\description{
Get the mlr Tuning Object
}
\seealso{
Other HyperControl: \code{\link{getHyperControlMeasures}},
  \code{\link{getHyperControlResampling}},
  \code{\link{makeHyperControl}},
  \code{\link{setHyperControlMeasures}},
  \code{\link{setHyperControlMlrControl}},
  \code{\link{setHyperControlResampling}}
}
