% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLearner_.R
\name{getLearnerClass}
\alias{getLearnerClass}
\title{Get the class of an mlr Learner}
\usage{
getLearnerClass(learner)
}
\arguments{
\item{learner}{[\code{Learner}]
Learner}
}
\value{
[\code{character(1)}]
}
\description{
The class consits of the type such as \dQuote{classif} or \dQuote{regr} and the name.
For example \dQuote{classif.svm}.
}
