% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLearner_.R
\name{getLearnerName}
\alias{getLearnerName}
\title{Get the name of an mlr Learner}
\usage{
getLearnerName(learner)
}
\arguments{
\item{learner}{[\code{Learner}]
Learner}
}
\value{
[\code{character(1)}]
}
\description{
The name describes the general \dQuote{learner} that is used without narrowing it down to  of the type such as \dQuote{classif} or \dQuote{regr} and the name.
For example the name \dQuote{classif.svm} is \dQuote{svm}.
}
