% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{getParConfigLearnerName}
\alias{getParConfigLearnerName}
\title{Get the name of the associated learner}
\usage{
getParConfigLearnerName(par.config)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}
}
\value{
[\code{character(1)}|NULL].
}
\description{
Get the name of the associated learner.
 Can be \code{NULL} when no learner is specified
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigNote}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
\concept{ParConfig}
