% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{getParConfigLearnerType}
\alias{getParConfigLearnerType}
\title{Get the type of the associated learner}
\usage{
getParConfigLearnerType(par.config)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}
}
\value{
[\code{character(1)}|NULL].
}
\description{
Get the type of the associated learner.
 Can be \code{NULL} when no type is specified
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
