% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{getParConfigNote}
\alias{getParConfigNote}
\title{Get the note}
\usage{
getParConfigNote(par.config)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}
}
\value{
[\code{ParConfig}].
}
\description{
Get the note attached to a Parameter Configuration.
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
