% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{getParConfigParVals}
\alias{getParConfigParVals}
\title{Get the constant parameter settings of the configuration}
\usage{
getParConfigParVals(par.config)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}
}
\value{
[\code{list}].
}
\description{
Get the constant parameter settings of the configuration
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigNote}},
  \code{\link{getParConfigParSet}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
