% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{makeHyperControl}
\alias{HyperControl}
\alias{makeHyperControl}
\title{Hyperparameter Tuning Control Object}
\usage{
makeHyperControl(mlr.control = NULL, resampling = NULL, measures = NULL)
}
\arguments{
\item{mlr.control}{[\code{\link[mlr]{TuneControl}}]\cr
Control object for search method. Also selects the optimization algorithm for tuning.}

\item{resampling}{[\code{\link[mlr]{ResampleDesc}}]
The resampling determines how the performance is obtained during tuning.}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}
}
\value{
[\code{HyperControl}]
}
\description{
Defines how the hyperparameter tuning should be conducted
}
\seealso{
Other HyperControl: \code{\link{getHyperControlMeasures}},
  \code{\link{getHyperControlMlrControl}},
  \code{\link{getHyperControlResampling}},
  \code{\link{setHyperControlMeasures}},
  \code{\link{setHyperControlMlrControl}},
  \code{\link{setHyperControlResampling}}
}

