% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{makeParConfig}
\alias{makeParConfig}
\alias{ParConfig}
\title{Make a Parameter Configuration}
\usage{
makeParConfig(par.set, learner = NULL, par.vals = NULL,
  learner.name = NULL, note = character(1L))
}
\arguments{
\item{par.set}{[\code{ParamSet}]
The Parameter Set.}

\item{learner}{[\code{Learner}|\code{character(1)}]
mlr Learner or string that identifies the Learner.}

\item{par.vals}{[\code{list}]
Specific constant parameter settings.
If the learner also has defined hyperparameters they will be combined using \code{link[BBmisc]{insert}} and a warning will be issued.}

\item{learner.name}{[\code{character(1)}]
Will be overwritten if \code{learner} is passed.
Can be used to associate the Parameter Configuration to a Learner without specifying if it belongs to e.g. classification or regression.}

\item{note}{[\code{character(1)}]
The note you want to attach to the Parameter Configuration.}
}
\value{
[\code{ParConfig}]
}
\description{
Defines a Parameter Configuration consisting of a \code{ParamSet} that defines the search range for the Hyperparameter Optimization.
Additionally you might want to specify some hard Parameters that are not supposed to be optimized but are necessary to be specified.
You might want to pass a Learner to associate the \code{ParConfig} to a special Learner.
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
