% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{setHyperControlMeasures}
\alias{setHyperControlMeasures}
\title{Set the measures}
\usage{
setHyperControlMeasures(hyper.control, measures)
}
\arguments{
\item{hyper.control}{[\code{\link{HyperControl}}]
The Hyperparameter Control Object}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}
}
\value{
[\code{HyperControl}]
}
\description{
Set the measures
}
\seealso{
Other HyperControl: \code{\link{getHyperControlMeasures}},
  \code{\link{getHyperControlMlrControl}},
  \code{\link{getHyperControlResampling}},
  \code{\link{makeHyperControl}},
  \code{\link{setHyperControlMlrControl}},
  \code{\link{setHyperControlResampling}}
}

