% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{setHyperControlMlrControl}
\alias{setHyperControlMlrControl}
\title{Set the mlr TuneControl Object}
\usage{
setHyperControlMlrControl(hyper.control, mlr.control)
}
\arguments{
\item{hyper.control}{[\code{\link{HyperControl}}]
The Hyperparameter Control Object}

\item{mlr.control}{[\code{\link[mlr]{TuneControl}}]\cr
Control object for search method. Also selects the optimization algorithm for tuning.}
}
\value{
[\code{HyperControl}]
}
\description{
Set the mlr TuneControl Object
}
\seealso{
Other HyperControl: \code{\link{getHyperControlMeasures}},
  \code{\link{getHyperControlMlrControl}},
  \code{\link{getHyperControlResampling}},
  \code{\link{makeHyperControl}},
  \code{\link{setHyperControlMeasures}},
  \code{\link{setHyperControlResampling}}
}

