% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{setHyperControlResampling}
\alias{setHyperControlResampling}
\title{Set the mlr resampling Object}
\usage{
setHyperControlResampling(hyper.control, resampling)
}
\arguments{
\item{hyper.control}{[\code{\link{HyperControl}}]
The Hyperparameter Control Object}

\item{resampling}{[\code{\link[mlr]{ResampleDesc}}]
The resampling determines how the performance is obtained during tuning.}
}
\value{
[\code{HyperControl}]
}
\description{
Set the mlr resampling Object
}
\seealso{
Other HyperControl: \code{\link{getHyperControlMeasures}},
  \code{\link{getHyperControlMlrControl}},
  \code{\link{getHyperControlResampling}},
  \code{\link{makeHyperControl}},
  \code{\link{setHyperControlMeasures}},
  \code{\link{setHyperControlMlrControl}}
}
