% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{setParConfigLearner}
\alias{setParConfigLearner}
\title{Set an associated Learner}
\usage{
setParConfigLearner(par.config, learner)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{learner}{[\code{Learner}]}
}
\value{
[\code{ParConfig}].
}
\description{
Associate the Param Configuration with a specific learner
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
