% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{setParConfigLearnerType}
\alias{setParConfigLearnerType}
\title{Set the type of the associated learner}
\usage{
setParConfigLearnerType(par.config, type)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{type}{[\code{character(1)}]}
}
\value{
[\code{ParConfig}].
}
\description{
Can be useful when you want to use Parameter Configuration on a different type of task.
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
