% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{setParConfigParSet}
\alias{setParConfigParSet}
\title{Set a new Parameter Set}
\usage{
setParConfigParSet(par.config, par.set)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{par.set}{[\code{ParamSet}]
The Parameter Set.}
}
\value{
[\code{ParConfig}].
}
\description{
Set a new Parameter Set
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigNote}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigNote}},
  \code{\link{setParConfigParVals}}
}
\concept{ParConfig}
