% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{setParConfigParVals}
\alias{setParConfigParVals}
\title{Set Parameter Values}
\usage{
setParConfigParVals(par.config, par.vals)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{par.vals}{[\code{list}]
Specific constant parameter settings.}
}
\value{
[\code{ParConfig}].
}
\description{
Set Parameter Values
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigParSet}}
}
