% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadParConfig.R
\name{uploadParConfig}
\alias{uploadParConfig}
\title{Uploads a Parameter Set to the mlrHyperopt servers}
\usage{
uploadParConfig(par.config, user.email = NULL)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{user.email}{[\code{character(1)}]
Your email to identify yourself to the server.
Does not have to be a valid one, but this identifier makes it easier to find your own submissions or submissions by colleagues.
In the future we might switch to an authentication system.
It could be advantageous to supply a working email if you want to migrate your submissions then.}
}
\value{
[\code{character}]
}
\description{
Uploads a ParConfig and returns the ID.
}
\examples{
par.set = makeParamSet(
  makeIntegerParam(
    id = "mtry",
    lower = expression(floor(p^0.25)),
    upper = expression(ceiling(p^0.75)),
    default = expression(round(p^0.5))),
  keys = "p")
par.config = makeParConfig(
  par.set = par.set,
  par.vals = list(ntree = 200),
  learner.name = "randomForest"
  )
\dontrun{
id = uploadParConfig(par.config, "jon.doe@example.com")
print(id)
}
}
