% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadParSet.R
\name{uploadParSet}
\alias{uploadParSet}
\title{Uploads a Parameter Set to the mlrHyperopt servers}
\usage{
uploadParSet(par.set, learner = NULL, user.email = NULL)
}
\arguments{
\item{par.set}{[\code{ParSet}]
The Parameter Set that you want to upload.}

\item{learner}{[\code{Learner}]
The Learner that belongs to that Parameter Set.
Can be left out if the \code{par.set} contains a reference to a learner.}

\item{user.email}{[\code{character(1)}]
Your email to identify yourself to the server.
Does not have to be a valid one, but this identifier makes it easier to find your own submissions or submissions by colleagues.
In the future we might switch to an authentification system.
It could be advantageous to supply a working email if you want to migrate your submissions then.}
}
\value{
[\code{character}]
}
\description{
Uploads a ParamSet and returns the ID.
}

