% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHyperControl.R
\name{generateHyperControl}
\alias{generateHyperControl}
\title{Generates a hyperparameter tuning control object}
\usage{
generateHyperControl(task, learner, par.config)
}
\arguments{
\item{task}{[\code{Task}]
Task}

\item{learner}{[\code{Learner}]
Learner}

\item{par.config}{[\code{ParConfig}]}
}
\value{
[\code{HyperControl}]
}
\description{
Tries to automatically create a suitable hyperparameter tuning control.
}
\examples{
par.config = getDefaultParConfig("regr.randomForest")
hyper.control = generateHyperControl(bh.task, "regr.randomForest", par.config)
# get what is inside
getHyperControlMeasures(hyper.control)
getHyperControlMlrControl(hyper.control)
getHyperControlResampling(hyper.control)
# change what is inside
hyper.control = setHyperControlMeasures(hyper.control, measures = medse)
hyper.control = setHyperControlMlrControl(
  hyper.control,
  mlr.control = makeTuneControlRandom(maxit = 10))
hyper.control = setHyperControlResampling(hyper.control, resampling = cv3)
hyperopt(task = bh.task, par.config = par.config, hyper.control = hyper.control)
}

