% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateParConfig.R
\name{generateParConfig}
\alias{generateParConfig}
\title{Gemerates a suitable Parameter Configuration Set for a given Task and a Learner}
\usage{
generateParConfig(task, learner)
}
\arguments{
\item{task}{[\code{Task}]
Task the learner should be tuned on.}

\item{learner}{[\code{Learner}]
The learner that is subject to the Hyperparameter Tuning.}
}
\value{
[\code{ParConfig}]
}
\description{
If now good Parameter Configuration Set is known this function tries to find one based on the given info.
}
\note{
The output of this function might change over time with coming updates. Not recommended for reprocuible research unless you save the output.
}
\examples{
par.config = generateParConfig(iris.task, "classif.svm")
print(par.config)
}

