% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultParConfig.R
\name{getDefaultParConfig}
\alias{getDefaultParConfig}
\title{Gives a default ParConfig for a Learner}
\usage{
getDefaultParConfig(learner)
}
\arguments{
\item{learner}{[\code{Learner}]
An mlr Learner.}
}
\value{
[\code{ParConfig}]
}
\description{
For a given Learner this method will return a useful default Parameter Set for Tuning.
}
\examples{
learner = makeLearner("classif.randomForest")
par.config = getDefaultParConfig(learner)
print(par.config)
}
