% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{getHyperControlResampling}
\alias{getHyperControlResampling}
\title{Get the Resample Description}
\usage{
getHyperControlResampling(hyper.control)
}
\arguments{
\item{hyper.control}{[\code{\link{HyperControl}}]
The Hyperparameter Control Object}
}
\value{
[\code{\link[mlr]{ResampleDesc}}|\code{\link[mlr]{ResampleInstance}}].
}
\description{
Get the Resample Description
}
\seealso{
Other HyperControl: 
\code{\link{getHyperControlMeasures}()},
\code{\link{getHyperControlMlrControl}()},
\code{\link{makeHyperControl}()},
\code{\link{setHyperControlMeasures}()},
\code{\link{setHyperControlMlrControl}()},
\code{\link{setHyperControlResampling}()}
}
\concept{HyperControl}
