% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperopt.R
\name{hyperopt}
\alias{hyperopt}
\title{Tune Hyperparameters for a machine learning task}
\usage{
hyperopt(task, learner = NULL, par.set = NULL, par.set.id = NULL,
  control = NULL)
}
\arguments{
\item{task}{[\code{Task}]
mlr Task}

\item{learner}{[\code{Learner}]
The learner that is subject to the Hyperparameter Tuning.
If no learner is given the learner referenced in the \code{par.set} will be used, if available.}

\item{par.set}{[\code{ParamSet}]
ParamHelpers Parameter Set}

\item{par.set.id}{[\code{character}]
Unique identifier for a Parameter Set.
This will be downloaded from the mlrHyperopt servers.}

\item{control}{[\code{list}]
This control object defines how the Hyperparameter Optimization should be conducted.}
}
\description{
Tunes the Hyperparameters for a given task and learner.
Tries to find the best parameter set to tune for the given learner.
}

