% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperControl.R
\name{makeHyperControl}
\alias{makeHyperControl}
\alias{HyperControl}
\title{Hyperparameter Tuning Control Object}
\usage{
makeHyperControl(mlr.control = NULL, resampling = NULL, measures = NULL,
  par.config = NULL)
}
\arguments{
\item{mlr.control}{[\code{\link[mlr]{TuneControl}}]\cr
Control object for search method. Also selects the optimization algorithm for tuning.}

\item{resampling}{[\code{\link[mlr]{ResampleDesc}}]
The resampling determines how the performance is obtained during tuning.}

\item{measures}{[\code{\link[mlr]{Measure}} | list of \code{\link[mlr]{Measure}}]\cr
Performance measure(s) to evaluate.
Default is the default measure for the task, see here \code{\link{getDefaultMeasure}}.}

\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{learner}{[\code{Learner}]
Learner}
}
\value{
[\code{HyperControl}]
}
\description{
Defines how the hyperparameter tuning should be conducted
}
\examples{
hyper.control = makeHyperControl(
  mlr.control = makeTuneControlRandom(maxit = 5),
  resampling = cv2,
  measures = acc
)
hyperopt(task = iris.task, learner = "classif.svm", hyper.control = hyper.control)
}
\seealso{
Other HyperControl: \code{\link{getHyperControlMeasures}},
  \code{\link{getHyperControlMlrControl}},
  \code{\link{getHyperControlResampling}},
  \code{\link{setHyperControlMeasures}},
  \code{\link{setHyperControlMlrControl}},
  \code{\link{setHyperControlResampling}}
}
