% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParConfig.R
\name{setParConfigNote}
\alias{setParConfigNote}
\title{Set a note}
\usage{
setParConfigNote(par.config, note)
}
\arguments{
\item{par.config}{[\code{\link{ParConfig}}]
The Parameter Configuration}

\item{note}{[\code{character(1)}]
The note.}
}
\value{
[\code{ParConfig}].
}
\description{
Attach note to the Parameter Configuration
}
\seealso{
Other ParConfig: \code{\link{getParConfigLearnerClass}},
  \code{\link{getParConfigLearnerName}},
  \code{\link{getParConfigLearnerType}},
  \code{\link{getParConfigNote}},
  \code{\link{getParConfigParSet}},
  \code{\link{getParConfigParVals}},
  \code{\link{makeParConfig}},
  \code{\link{setParConfigLearnerType}},
  \code{\link{setParConfigLearner}},
  \code{\link{setParConfigParSet}},
  \code{\link{setParConfigParVals}}
}
\concept{ParConfig}
