% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadParSet.R
\name{uploadParSet}
\alias{uploadParSet}
\title{Uploads a Parameter Set to the mlrHyperopt servers}
\usage{
uploadParSet(par.set, learner = NULL, user.email = NULL)
}
\arguments{
\item{id}{[\code{character}]
A unique identifier of the Parameter Set}
}
\value{
[\code{character}]
}
\description{
Uploads a ParamSet and returns the ID.
}

